requirements.txt
alembic==1.7.3
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
six==1.16.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.8
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.5.0
